requirements.txt
# Core dependencies - Genius Trading System
pybit==5.6.2
python-dotenv==1.0.0
numpy==1.26.3
pytz==2024.1

# Optional for monitoring
pandas==2.1.4